requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
arn==0.1.5
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
boto3==1.26.22
botocore==1.29.22
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==0.5.0
Django==4.1
django-braces==1.15.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-model-utils==4.2.0
django-star-ratings==0.9.2
django-storages==1.13.1
djangoajax==3.3
document==1.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
pep8-naming==0.13.2
Pillow==9.3.0
platformdirs==2.5.3
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
s3transfer==0.6.0
self==2020.12.3
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
sweetify==2.3.1
tomli==2.0.1
tzdata==2022.5
urllib3==1.26.13
whitenoise==6.2.0
yarl==1.8.1
